update and test ResourceUseController

+++ app/Http/Controllers/ResourceUseController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\ResourceUse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 
 class ResourceUseController extends Controller
 {
@@ -13,7 +16,7 @@ class ResourceUseController extends Controller
      */
     public function index()
     {
-        //
+        return view('resource_uses');
     }
 
     /**
@@ -34,7 +37,16 @@ class ResourceUseController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $this->authorize('create', ResourceUse::class);
+		// first, validate the request
+		$validated = $request->validate([
+			'name' => 'string|required|max:255'
+		]);
+
+		// create a new ResourceUse using mass assignment to add the 'name' attribute
+		$resource_use = (new ResourceUse)->fill($validated);
+		$resource_use->author_id = Auth::id();
+		$resource_use->save();
     }
 
     /**
@@ -66,10 +78,10 @@ class ResourceUseController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
-    {
-        //
-    }
+    // public function update(Request $request, $id)
+    // {
+    //     //
+    // }
 
     /**
      * Remove the specified resource use from storage.
@@ -79,6 +91,7 @@ class ResourceUseController extends Controller
      */
     public function destroy(ResourceUse $resource_use)
     {
+        $this->authorize('delete', ResourceUse::class);
         $resource_use->delete();
     }
 }
